+++ Realm/Tests/MigrationTests.mm
@@ -351,31 +351,36 @@ extern "C" {
     [RLMRealm migrateRealmAtPath:RLMTestRealmPath()];
 }
 
-- (void)testStringPrimaryKeyNoPKTableMigration {
+- (void)testIntPrimaryKeyNoIndexMigration {
     // make string an int
-    RLMObjectSchema *objectSchema = [RLMObjectSchema schemaForObjectClass:MigrationStringPrimaryKeyObject.class];
-    objectSchema.primaryKeyProperty.isPrimary = NO;
-    objectSchema.primaryKeyProperty = nil;
+    RLMObjectSchema *objectSchema = [RLMObjectSchema schemaForObjectClass:MigrationPrimaryKeyObject.class];
+
+    // create without search index
+    objectSchema.primaryKeyProperty.attributes = 0;
 
     // create realm with old schema and populate
-    RLMRealm *realm = [self realmWithSingleObject:objectSchema];
-    [realm beginWriteTransaction];
-    [realm createObject:MigrationStringPrimaryKeyObject.className withObject:@[@"1"]];
-    [realm createObject:MigrationStringPrimaryKeyObject.className withObject:@[@"2"]];
+    @autoreleasepool {
+        RLMRealm *realm = [self realmWithSingleObject:objectSchema];
+        [realm beginWriteTransaction];
+        [realm createObject:MigrationPrimaryKeyObject.className withObject:@[@1]];
+        [realm createObject:MigrationPrimaryKeyObject.className withObject:@[@2]];
+        [realm commitWriteTransaction];
 
-    // remove the PK table to match old versions that create it only when needed
-    realm.group->remove_table(c_primaryKeyTableName);
-    [realm commitWriteTransaction];
+        XCTAssertFalse(realm.schema[MigrationPrimaryKeyObject.className].table->has_search_index(0));
+    }
 
     // apply migration
-    [RLMRealm setSchemaVersion:1
-                forRealmAtPath:RLMTestRealmPath()
-            withMigrationBlock:^(__unused RLMMigration *migration, __unused NSUInteger oldSchemaVersion) {
-        [migration enumerateObjects:@"MigrationStringPrimaryKeyObject" block:^(__unused RLMObject *oldObject, RLMObject *newObject) {
-            newObject[@"stringCol"] = [[NSUUID UUID] UUIDString];
-        }];
-    }];
-    [self realmWithTestPath];
+    [RLMRealm setSchemaVersion:1 forRealmAtPath:RLMTestRealmPath() withMigrationBlock:^(__unused RLMMigration *migration, __unused NSUInteger oldSchemaVersion) { }];
+    [RLMRealm migrateRealmAtPath:RLMTestRealmPath()];
+
+    // check that column is now indexed
+    RLMRealm *realm = [self realmWithTestPath];
+    XCTAssertTrue(realm.schema[MigrationPrimaryKeyObject.className].table->has_search_index(0));
+
+    // verify that old data still exists
+    RLMResults *objects = [MigrationPrimaryKeyObject allObjectsInRealm:realm];
+    XCTAssertEqual(1, [objects[0] intCol]);
+    XCTAssertEqual(2, [objects[1] intCol]);
 }
 
 - (void)testDuplicatePrimaryKeyMigration {