+++ CHANGELOG.md
@@ -12,6 +12,8 @@ x.x.x Release notes (yyyy-MM-dd)
 ### Enhancements
 
 * Add `-[RLMRealm writeCopyToPath:encryptionKey:error:]`.
+* Int properties can now be indexed, and int primary keys are automatically
+  indexed (as string primary keys already were).
 
 ### Bugfixes
 
@@ -20,8 +22,8 @@ x.x.x Release notes (yyyy-MM-dd)
   while it's being fast-enumerated.
 * Also encrypt the temporary files used when encryption is enabled for a Realm.
 * Fixed crash in JSONImport example on OS X with non-en_US locale.
-* Fix a crash when adding primary keys to older realm files with no primary
-  keys on any objects.
+* Fixed infinite loop when opening a Realm file in the Browser at the same time
+  as it is open in a 32-bit simulator.
 
 0.89.2 Release notes (2015-01-02)
 =============================================================